Move exceptions to their own package.
