+++ src/main/java/odin/exceptions/IllegalFileException.java
@@ -1,4 +1,4 @@
-package odin;
+package odin.exceptions;
 
 public class IllegalFileException extends RuntimeException {
     public IllegalFileException(String message) {
