+++ src/main/java/odin/exceptions/IllegalTaskException.java
@@ -1,4 +1,4 @@
-package odin;
+package odin.exceptions;
 
 public class IllegalTaskException extends Exception {
     public IllegalTaskException(String message) {
